Move ftml files to a subdirectory and split out data classes.
